trying to fix FTP/HTTP methods and decouple netstart support
Propagator's FTP/HTTP methods should download and check squashfs instead
of full ISO-image. In contrast, Netstart requires a full ISO-image. Now
this is toggled with the ramdisk_size argument: when it is defined the
ISO-image should already be mounted on the LAN server and altboot will
only download the stage2 squashfs file from the mounted directory,
and if ramdisk_size is not defined then the full ISO-image will be
downloaded from the internet server, the hash will be checked against
the full ISO-image instead of squashfs, and the standard HTTP/FTP
variables will not be exported to the stage2.