squash parameter introduced, separated from fuid
The fuid parameter is no longer used internally, the squash parameter
has been added for this purpose. Now fuid is also used as a "soft UUID"
that gives a hint about the original ISO image when scanning disks.
Minor errors in the bootchain-localdev module and in the documentation
have been fixed, the ability to boot from an image unpacked into
a subdirectory has been added. The documentation has been updated
to the current version.

See also: https://bugzilla.altlinux.org/54071